Initial Verona Dialect commit
From the standalone example, merging the CMake files and tests with the
existing Verona build infrastructure. Only builds with the CMake flag
-DEXPERIMENTAL_MLIR_COMPILER=True, as well as the LLVM build directory
path -DLLVM_BUILD=/abs/path/to/llvm-project/build.

Added the `dummy.verona` test to check on the basic validity of the
conversion tool `verona-opt`, which will take MLIR->MLIR and potentially
optimise it.

However, the main task here will be the AST->MLIR, a different tool,
using the ast and mlit libraries to convert one to another. The actual
compiler, later on, will use all of the libraries in the same way, but
not call these tools directly.

This way, we have individual tools for testing as well as guaranteeing a
more generic API for all libraries.