Adding IntegerType + storage, removes OpaqueType
This commit adds a new base type, IntegerType, with parameters `width`
and `sign`, that can create different types. This is a proof of concept,
and an example for future types to come.

We don't need an opaque type, we just need to allow unknown types and
call them anything other than !verona<"...">.

Changing the dummy test to match a specific type (!verona.U64) as well
as an opaque type (!type<"U64 & imm">) and test the new type system.