Parsing function body, arguments, return values
The current generator doesn't handle types at all, they're all opaque
and string based. There is a cast at the return to the function's return
type, as we're relying on the ast to have done the minimal checks, and
on the subsequent (to be implemented) type check to make sure it's
correct.

The operations are also opaque for now, function-like, returning single
values. This allows us to create unary and binary operators with varied
syntax ("verona.add", "verona.cast", etc). This flexibility helps us
generate code without caring much about the structure, which was checked
by the AST generator and will be validated by the lowering to Verona
dialect.

Additional tasks: Reimplemented symbol table locally

The ScopedHashTable works well for symbols that cannot be updated, but
it doesn't allow us to redefine the value of symbols. This is bad
because we want to declare variables and later bind values to them.

This is a simplistic approach that does the same thing, but allows later
binding.